WGLC (#16)
* Prepare for draft-17, WGLC

* RTCWeb -> WebRTC, MSRP -> msrp, m -> m=, IESG as contact for IANA (#12)

* 'for the' nits, IANA subproto registration removed, tranport ABNF, table size (#13)

* Update examples, MSRP -> msrp

* 'for the' nits, IANA subproto registration removed, tranport ABNF, table size

* Updates after WGLC comments round (#14)

Session closure, data receiving
Draft now states that it updates RFC4975 and adds an IANA consideration to update the msrps registration, to include DTLS
Removed RFC4145 and other dangling references
Introduction corrected "MSRP is currently defined ... " -> "MSRP was initially defined in ..."
Added text to make more clear the scope of path and MSRP URI:
Data channels are negotiated independently of the value of the "path" attribute. This means that when MSRP messages are transported on a data channel, the "path" attribute is not used for transport negotiation or routing of the messages. MSRP endpoints using data channels can set the value of the "path" attribute, including the MSRP URI, and use it as described in RFC4975
Simplified SDP session closing section, referring back to mmusic-data-channel-sdpneg
Updated examples removing the port
Updated introduction to MSRP considerations section "This section describes the MSRP ... " -> "The procedures specified in RFC4975 apply except ..."
Clarification on MSRP session closing section, explicitly stating that MSRP session should be considered failed when data channels are torn down

* Additional WGLC (#15)

* MUST -> SHALL; revert port removal in examples; rewrite wording on path and transport negotiation

* Remove links for purged references in change log; reorder updates, security and IANA sections for consistency

* m/c alignment in example; path clarification; introduction clarification

* Typos, acknowledgements

* Typos